// frontend/src/components/receiptVoucher/types.ts
export type ReceiptVoucherSearchQueries = {
  year: number | string;
  month: number | string;
  day: number | string;
  customerName: string;
};

export type FilterProps = {
  setIsFilterOpen: React.Dispatch<React.SetStateAction<boolean>>;
  isFilterOpen: boolean;
  setSearchQuery: React.Dispatch<
    React.SetStateAction<ReceiptVoucherSearchQueries>
  >;
  searchQuery: ReceiptVoucherSearchQueries;
  setTableRows: React.Dispatch<React.SetStateAction<number>>;
  tableRows: number;
};

export type FilterSummaryProps = {
  searchQuery: ReceiptVoucherSearchQueries;
  count: number;
  setIsFilterOpen: React.Dispatch<React.SetStateAction<boolean>>;
  isFilterOpen: boolean;
};